+++ Classes/ViewHelpers/PriceViewHelper.php
@@ -8,15 +8,15 @@ use OliverKlee\Oelib\Exception\NotFoundException;
 use OliverKlee\Oelib\Mapper\CurrencyMapper;
 use OliverKlee\Oelib\Mapper\MapperRegistry;
 use OliverKlee\Oelib\Model\Currency;
-use TYPO3\CMS\Fluid\Core\ViewHelper\AbstractViewHelper;
+use TYPO3Fluid\Fluid\Core\ViewHelper\AbstractViewHelper;
 
 /**
  * This class represents a view helper for formatting a price.
  *
- * The value (setValue()) and the currency (setCurrencyFromIsoAlpha3Code())
- * should be set before calling render(). You can use the same instance of this
+ * The value (`setValue()`) and the currency (`setCurrencyFromIsoAlpha3Code()`)
+ * should be set before calling `render()`. You can use the same instance of this
  * view helper to render different values in the same currency by changing the
- * value via setValue().
+ * value via `setValue()`.
  */
 class PriceViewHelper extends AbstractViewHelper
 {
@@ -44,7 +44,7 @@ class PriceViewHelper extends AbstractViewHelper
      * Sets the currency of the price to render based on the currency's ISO
      * alpha 3 code, e.g. "EUR" for Euro, "USD" for US dollars.
      *
-     * @param string$isoAlpha3Code the ISO alpha 3 code of the currency to set, must not be empty
+     * @param string $isoAlpha3Code the ISO alpha 3 code of the currency to set, must not be empty
      */
     public function setCurrencyFromIsoAlpha3Code(string $isoAlpha3Code): void
     {